refactor campaign route
