+++ src/features/routes/UserRoute.ts
@@ -45,6 +45,11 @@ export default class UserRoute<T extends RouteClassTypes> extends Route<T> {
     if (!olp) return false;
     return olp === true || olp?.includes(campaignId);
   }
+  protected hasAccessToProspect(campaignId: number) {
+    const olp = this.configuration.request.user.permission.admin?.appq_prospect;
+    if (!olp) return false;
+    return olp === true || olp?.includes(campaignId);
+  }
 
   protected hasAccessTesterSelection(campaignId: number) {
     const olp =
