+++ src/routes/campaigns/campaignId/stats/_get/index.ts
@@ -1,11 +1,23 @@
 /** OPENAPI-CLASS: get-campaigns-campaign-stats */
-import AdminCampaignRoute from "@src/features/routes/AdminCampaignRoute";
+import AdminCampaignRoute from "@src/features/routes/CampaignRoute";
 import { tryber } from "@src/features/database";
+import OpenapiError from "@src/features/OpenapiError";
 
 export default class StatsRoute extends AdminCampaignRoute<{
   response: StoplightOperations["get-campaigns-campaign-stats"]["responses"]["200"]["content"]["application/json"];
   parameters: StoplightOperations["get-campaigns-campaign-stats"]["parameters"]["path"];
 }> {
+  protected async filter(): Promise<boolean> {
+    if (!(await super.filter())) return false;
+
+    if (!this.hasAccessToBugs(this.cp_id)) {
+      this.setError(403, new OpenapiError("Access denied"));
+
+      return false;
+    }
+    return true;
+  }
+
   protected async prepare(): Promise<void> {
     let stats;
     try {
